% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance_between_trajectories.R
\name{calculate_distance_between_trajectories}
\alias{calculate_distance_between_trajectories}
\title{Calculate distance between trajectories}
\usage{
calculate_distance_between_trajectories(
  microbiota_age_data,
  filter_expression = TRUE,
  min_threshold = 0.1,
  groups_to_compare = list(`CG-HMG` = c("CG", "HMG"), `TG1-HMG` = c("TG1", "HMG"),
    `TG2-HMG` = c("TG2", "HMG")),
  xlab = "Chronological age (days)",
  ylab = "Distance between trajectories",
  save_figure = FALSE,
  figure_name = "trajectory_distance.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
