% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MAZ.R
\name{calculate_MAZ}
\alias{calculate_MAZ}
\title{Calculate Microbiota-for-age Z-scores (MAZ)}
\usage{
calculate_MAZ(
  microbiota_age_data,
  reference_expression = expression(Feeding_group == "HMG")
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
