% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_biome_between_on_off_trajectory.R
\name{compare_biome_between_on_off_trajectory}
\alias{compare_biome_between_on_off_trajectory}
\title{Compare microbiome or biomarkers between on- and off-trajectory subjects}
\usage{
compare_biome_between_on_off_trajectory(
  biome_data,
  on_off_trajectory_data,
  filter_expression = expression(Feeding_group != "HMG"),
  loop_over_visits = c("3 months", "6 months", "12 months", "15 months"),
  FDR_threshold = 0.05
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
