% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_on_off_trajectory_between_groups.R
\name{compare_on_off_trajectory_between_groups}
\alias{compare_on_off_trajectory_between_groups}
\title{Compare feeding groups between on- and off-trajectory subjects}
\usage{
compare_on_off_trajectory_between_groups(
  on_off_trajectory_data,
  filter_expression = TRUE,
  TG_group,
  trajectory_status_variable,
  trend = "CG-TG-HMG"
)
}
\value{
Documentation will follow
}
\description{
Using Cochran-Armitage trend test
Documentation will follow
}
