% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/compare_on_off_trajectory_between_groups_all_combinations.R
\name{compare_on_off_trajectory_between_groups_all_combinations}
\alias{compare_on_off_trajectory_between_groups_all_combinations}
\title{Compare feeding groups between on- and off-trajectory subjects - all combinations used in the study}
\usage{
compare_on_off_trajectory_between_groups_all_combinations(
  on_off_trajectory_data,
  filter_expression = TRUE
)
}
\value{
Documentation will follow
}
\description{
Using Cochran-Armitage trend test
Documentation will follow
}
