% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trajectories_between_groups.R
\name{compare_trajectories_between_groups}
\alias{compare_trajectories_between_groups}
\title{Compare trajectories between feeding groups}
\usage{
compare_trajectories_between_groups(
  microbiota_age_data,
  filter_expression = TRUE,
  interval = 60,
  groups_to_compare = list(`CG-HMG` = c("CG", "HMG"), `TG1-HMG` = c("TG1", "HMG"),
    `TG2-HMG` = c("TG2", "HMG")),
  xlab = "Chronological age (days)",
  ylab = "P-value",
  threshold_point_color = "red",
  save_figure = FALSE,
  figure_name = "trajectory_comparison_pvalues.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
