% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_microbiome_biomarkers.R
\name{correlate_microbiome_biomarkers}
\alias{correlate_microbiome_biomarkers}
\title{Correlate microbiome data and biomarkers data}
\usage{
correlate_microbiome_biomarkers(
  microbiome_data,
  biomarker_data,
  metadata,
  filter_expression = expression(Feeding_group != "HMG"),
  filter_visit_expression = expression(Visit \%in\% c("3 months", "6 months",
    "12 months", "15 months")),
  FDR_threshold = 0.05
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
