% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom_theme}
\alias{custom_theme}
\title{Custom theme for plots}
\usage{
custom_theme(
  font_size = 10,
  line_size = 0.5,
  smaller_fraction = 0.8,
  axis_margin_fraction = 0.25
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
