% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_on_off_trajectory.R
\name{define_on_off_trajectory}
\alias{define_on_off_trajectory}
\title{Define on- and off-trajectory subjects}
\usage{
define_on_off_trajectory(
  maz_data,
  threshold = 3,
  baseline_visit = c("Baseline"),
  early_visits = c("3 months", "6 months"),
  late_visits = c("12 months", "15 months")
)
}
\value{
Documentation will follow
}
\description{
Based on Microbiota-for-age Z-scores
Documentation will follow
}
