% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_cox_ph_test.R
\name{perform_cox_ph_test}
\alias{perform_cox_ph_test}
\title{Perform Cox Proportional-Hazards test}
\usage{
perform_cox_ph_test(
  event_data,
  on_off_trajectory_data,
  filter_expression = expression(Feeding_group != "HMG"),
  trajectory_status_variable = "Trajectory_status_per_subject",
  time_limits = c(1 * 30, 15 * 30),
  covariates = c("Feeding_group", "Delivery_mode", "Sex")
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
