% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_cox_ph_test_all_combinations.R
\name{perform_cox_ph_test_all_combinations}
\alias{perform_cox_ph_test_all_combinations}
\title{Perform Cox Proportional-Hazards test - all combinations used in the study}
\usage{
perform_cox_ph_test_all_combinations(
  event_data,
  on_off_trajectory_data,
  filter_expression = expression(Feeding_group != "HMG")
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
