% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biomarker_on_off_trajectory.R
\name{plot_biomarker_on_off_trajectory}
\alias{plot_biomarker_on_off_trajectory}
\title{Plot biomarker measurements for on-/off-trajectory groups and each follow-up visit}
\usage{
plot_biomarker_on_off_trajectory(
  biomarker,
  biomarker_data,
  on_off_trajectory_data,
  filter_expression = expression(Feeding_group != "HMG"),
  filter_visit_expression = expression(Visit \%in\% c("3 months", "6 months",
    "12 months", "15 months")),
  add_units_string = TRUE,
  color_on_trajectory = "#30A190",
  color_off_trajectory = "#E09B65",
  label_on_trajectory = "On-trajectory",
  label_off_trajectory = "Off-trajectory",
  visit_levels = c("Baseline", "3 months", "6 months", "12 months", "15 months"),
  plot_title = "",
  plot_caption = "",
  title_fontsize = 12,
  subtitle_fontsize = 12,
  text_fontsize = 10,
  save_figure = FALSE,
  figure_name = glue::glue("{biomarker}_on_off_trajectory_boxplot.pdf")
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
