% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest_plot_cox_ph.R
\name{plot_forest_plot_cox_ph}
\alias{plot_forest_plot_cox_ph}
\title{Plot forest plot showing Cox PH results}
\usage{
plot_forest_plot_cox_ph(
  results_data,
  plot_title = NULL,
  adjusted_model_string = "Adjusted for Feeding_group, Delivery_mode, Sex",
  unadjusted_model_string = "Unadjusted",
  save_figure = FALSE,
  figure_name = "forest_plot_cox_ph.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
