% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kaplan_meier.R
\name{plot_kaplan_meier}
\alias{plot_kaplan_meier}
\title{Plot Kaplan-Meier curves}
\usage{
plot_kaplan_meier(
  event_data,
  on_off_trajectory_data,
  filter_expression = expression(Feeding_group != "HMG"),
  trajectory_status_variable = "Trajectory_status_per_subject",
  time_limits = c(1 * 30, 15 * 30),
  breaks = 30,
  caption_log_rank_test_results = TRUE,
  xlabel = "Months of age",
  ylabel = "Event",
  plot_title = "",
  plot_subtitle = "",
  legend_title = "",
  label_on_trajectory = "On-trajectory",
  label_off_trajectory = "Off-trajectory",
  color_on_trajectory = "#30A190",
  color_off_trajectory = "#E09B65",
  plot_title_fontsize = 16,
  plot_subtitle_fontsize = 14,
  axis_title_fontsize = 14,
  axis_text_fontsize = 14,
  legend_title_fontsize = 14,
  legend_text_fontsize = 14,
  cumevents_fontsize = 5,
  plot_caption_fontsize = 12,
  save_figure = FALSE,
  figure_name = "kaplan_meier_curve.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
