% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kaplan_meier_all_combinations.R
\name{plot_kaplan_meier_all_combinations}
\alias{plot_kaplan_meier_all_combinations}
\title{Plot Kaplan-Meier curves - all combinations used in the study}
\usage{
plot_kaplan_meier_all_combinations(
  event_data,
  event_name,
  on_off_trajectory_data,
  filter_expression = expression(Feeding_group != "HMG")
)
}
\value{
Documentation will follow
}
\description{
Output does not include table of cumulative events, only the KM curve.
Documentation will follow
}
