% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maz.R
\name{plot_maz}
\alias{plot_maz}
\title{Plot Microbiota-for-age Z-scores (MAZ) for each feeding group and visit}
\usage{
plot_maz(
  maz_data,
  filter_expression = TRUE,
  outlier_threshold = 3,
  color_values = list(CG = "#02AE71", TG1 = "#06AEFA", TG2 = "#3B6BC0", HMG = "#FFB829"),
  visit_levels = c("Baseline", "3 months", "6 months", "12 months", "15 months"),
  xlabel = "Feeding group",
  ylabel = "Microbiota-for-age Z-score",
  plot_title = NULL,
  save_figure = FALSE,
  figure_name = "maz_scores.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
