% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_on_off_trajectory_between_groups.R
\name{plot_on_off_trajectory_between_groups}
\alias{plot_on_off_trajectory_between_groups}
\title{Plot on-/off-trajectory per feeding group}
\usage{
plot_on_off_trajectory_between_groups(
  on_off_trajectory_data,
  filter_expression = TRUE,
  TG_group,
  trajectory_status_variable,
  trend = "CG-TG-HMG",
  show_n_barplots = FALSE,
  save_figure = FALSE,
  figure_name = "on_off_trajectory_groups_barplot.pdf",
  color_on_trajectory = "#30A190",
  color_off_trajectory = "#E09B65",
  label_on_trajectory = "On-trajectory",
  label_off_trajectory = "Off-trajectory",
  xlabel = "Feeding group",
  ylabel = "Proportion of subjects",
  y_axis_ticks_interval = 0.1,
  plot_title = "",
  plot_caption = NULL,
  plot_font_size = 14,
  legend_title = "",
  legend_title_size = 14,
  legend_text_size = 14,
  legend_key_size = 14
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
