% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reference_trajectory.R
\name{plot_reference_trajectory}
\alias{plot_reference_trajectory}
\title{Plot microbiota age trajectory from reference/training set}
\usage{
plot_reference_trajectory(
  microbiota_age_data,
  color = "#FFB829",
  xlabel = "Chronological age (months)",
  ylabel = "Microbiota age (months)",
  title = "",
  font_size = 12,
  save_figure = FALSE,
  figure_name = "microbiota_age_trajectory_reference.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
