% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap_values.R
\name{plot_shap_values}
\alias{plot_shap_values}
\title{Plot SHAP values (feature importance) from age predictor model}
\usage{
plot_shap_values(
  shap_data,
  model_performance_data,
  model,
  color = "#30A190",
  xlabel = "SHAP feature importance",
  limits = c(0, 70),
  title = "Model feature importance",
  feature_fontsize = 12,
  xlabel_size = 9,
  xtext_size = 8,
  save_figure = FALSE,
  figure_name = "shap_feature_importance.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
