% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot microbiota trajectory curves}
\usage{
plot_trajectories(
  microbiota_age_data,
  filter_expression = TRUE,
  ci = TRUE,
  alpha = 0.15,
  breaks = c(0, 3, 6, 9, 12, 15),
  color_values = list(CG = "#02AE71", TG1 = "#06AEFA", TG2 = "#3B6BC0", HMG = "#FFB829"),
  xlabel = "Chronological age (months)",
  ylabel = "Microbiota age (months)",
  color_label = "Feeding group",
  caption = "",
  save_figure = FALSE,
  figure_name = "microbiota_trajectories.pdf"
)
}
\value{
Documentation will follow
}
\description{
Documentation will follow
}
